+++ adapter_test.go
@@ -1,4 +1,4 @@
-package joe
+package joe_test
 
 import (
 	"bytes"
@@ -6,14 +6,16 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/go-joe/joe"
+	"github.com/go-joe/joe/joetest"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zaptest"
 )
 
-func cliTestAdapter(t *testing.T) (a *CLIAdapter, output *bytes.Buffer) {
+func cliTestAdapter(t *testing.T) (a *joe.CLIAdapter, output *bytes.Buffer) {
 	logger := zaptest.NewLogger(t)
-	a = NewCLIAdapter("test", logger)
+	a = joe.NewCLIAdapter("test", logger)
 	output = new(bytes.Buffer)
 	a.Output = output
 	return a, output
@@ -23,28 +25,23 @@ func TestCLIAdapter_Register(t *testing.T) {
 	input := new(bytes.Buffer)
 	a, output := cliTestAdapter(t)
 	a.Input = ioutil.NopCloser(input)
-	brain := NewBrain(a.Logger)
+	brain := joetest.NewBrain(t)
+	messages := brain.Events()
 
 	input.WriteString("Hello\n")
 	input.WriteString("World\n")
 
-	messages := make(chan ReceiveMessageEvent, 2)
-	brain.RegisterHandler(func(msg ReceiveMessageEvent) {
-		messages <- msg
-	})
-
-	a.RegisterAt(brain)
-
-	go brain.HandleEvents()
+	// Start the Goroutine of the adapter which consumes the input
+	a.RegisterAt(brain.Brain)
 
 	msg1 := <-messages
 	msg2 := <-messages
 
-	assert.Equal(t, "Hello", msg1.Text)
-	assert.Equal(t, "World", msg2.Text)
+	assert.Equal(t, "Hello", msg1.Data.(joe.ReceiveMessageEvent).Text)
+	assert.Equal(t, "World", msg2.Data.(joe.ReceiveMessageEvent).Text)
 
 	// Stop the brain to make sure we are done with all callbacks
-	brain.Shutdown(ctx)
+	brain.Finish()
 
 	// Close the adapter to finish up the test
 	assert.NoError(t, a.Close())
@@ -69,7 +66,7 @@ func TestCLIAdapter_Close(t *testing.T) {
 	input := new(bytes.Buffer)
 	a, output := cliTestAdapter(t)
 	a.Input = ioutil.NopCloser(input)
-	brain := NewBrain(a.Logger)
+	brain := joe.NewBrain(a.Logger)
 	a.RegisterAt(brain)
 
 	err := a.Close()